[Refactor] Use explicit type casting for narrow conversions
CWE-190, CWE-192, CWE-197, CWE-681.

+++ app/src/main/java/io/github/muntashirakon/io/SplitInputStream.java
@@ -171,7 +171,7 @@ public class SplitInputStream extends InputStream {
                     if (mMarkBuf != null) {
                         // Accumulate data in markBuf
                         System.arraycopy(mBuf, mMarkPos, mMarkBuf, mMarkBufCount, (int) size);
-                        mMarkBufCount += size;
+                        mMarkBufCount += (int) size;
                         // Set markPos to 0 as buffer will refill
                         mMarkPos = 0;
                     }
